Add eventwizard2 to py3
