+++ src/backend/web/templates/react-eventwizard.html
